update naming conventions
