+++ .README/README.md
@@ -54,7 +54,7 @@ npm install eslint babel-eslint eslint-plugin-flowtype --save-dev
       2,
       "never"
     ],
-    "flowtype/interface-type-id-match": [
+    "flowtype/interface-id-match": [
       2,
       "^([A-Z][a-z0-9]+)+Type$"
     ],
@@ -166,7 +166,7 @@ When `true`, only checks files with a [`@flow` annotation](http://flowtype.org/d
 {"gitdown": "include", "file": "./rules/delimiter-dangle.md"}
 {"gitdown": "include", "file": "./rules/enforce-line-break.md"}
 {"gitdown": "include", "file": "./rules/generic-spacing.md"}
-{"gitdown": "include", "file": "./rules/interface-type-id-match.md"}
+{"gitdown": "include", "file": "./rules/interface-id-match.md"}
 {"gitdown": "include", "file": "./rules/newline-after-flow-annotation.md"}
 {"gitdown": "include", "file": "./rules/no-dupe-keys.md"}
 {"gitdown": "include", "file": "./rules/no-existential-type.md"}
